Title: 괄호 쌍 만들어주기, Time: 131ms, Memory: 11MB, Status: Passed - CodeTree

+++ 240418/괄호 쌍 만들어주기/pair-parentheses.java
@@ -8,7 +8,7 @@ public class Main {
 
     private static int[] right;
 
-    private static int ans = 0;
+    private static long ans = 0;
 
     public static void main(String[] args) throws IOException {
         parentheses = br.readLine().toCharArray();
